Get filtered facilites also when updating list status, so we can get them when adding or removing to a table
